+++ lib/src/carvings/string/append.dart
@@ -2,7 +2,10 @@ import '../../interface.dart';
 
 class CarvingAppend extends Carving<String, String> {
 	final String replacement;
-	CarvingAppend(this.replacement);
+  final int? at;
+	CarvingAppend(this.replacement, { this.at });
 	
-	@override String apply(String input) => input + replacement;
+	@override String apply(String input) => at == null
+    ? input + replacement
+    : input.replaceRange(at!, at!, replacement);
 }